Logic Changes On SessionsToPerform

+++ src/main/java/com/restproject/backend/entities/SessionsOfSchedules.java
@@ -30,5 +30,5 @@ public class SessionsOfSchedules {
     Schedule schedule;
 
     @Column(name = "ordinal", nullable = false)
-    Long ordinal;
+    Integer ordinal;
 }
